requirements.txt
# Ingestion Pipeline Dependencies

# PDF Processing
PyMuPDF>=1.23.0

# Database
psycopg[binary]>=3.1.0

# YAML parsing
PyYAML>=6.0

# Excel processing (for coverage mapping loader)
pandas>=2.0.0
openpyxl>=3.1.0

# Optional: OpenAI Embedding (not required for dummy embedding)
# openai>=1.0.0
